code review (#11)
now you can get statistic not hard-code!!!
yhis is really statistic

+++ statistic-service/src/main/java/com/student/statisticservice/dto/StatisticDto.java
@@ -12,5 +12,5 @@ public class StatisticDto {
     String studentUuid;
     String studentName;
     Double grade;
-    Integer articlesCount;
+    Long articlesCount;
 }
